no-old-shims: Extract util function
