+++ lib/utils/new-module.js
@@ -38,6 +38,89 @@ function buildMessage(obj) {
   return message;
 }
 
+/**
+ * This function returns a fix function. The fix function can update the old
+ * imports to the new ones.
+ * @param {Object} node - the AST node related to the import problem
+ * @param {Object} modulesData - the data set that maps old imports to the new
+ * ones. It is an object like this one:
+ * {
+ *   'ember-array': {
+ *     default: ['@ember/array'],
+ *   },
+ *   'ember-array/mutable': {
+ *     default: ['@ember/array/mutable'],
+ *   },
+ *   'ember-array/utils': {
+ *     A: ['@ember/array', 'A'],
+ *     isEmberArray: ['@ember/array', 'isArray'],
+ *     wrap: ['@ember/array', 'makeArray'],
+ *   },
+ * }
+ * @return {function(fixer: Object): function} a function that applies a fix to
+ * update the import
+ */
+function buildFix(node, modulesData) {
+  const moduleName = node.source.value;
+  const moduleMappings = modulesData[moduleName];
+
+  const fix = function(fixer) {
+    const newImports = {};
+
+    node.specifiers.forEach(specifier => {
+      const localName = specifier.local.name;
+
+      let importedName;
+      if (specifier.type === 'ImportDefaultSpecifier') {
+        importedName = 'default';
+      } else {
+        importedName = specifier.imported.name;
+      }
+
+      let module;
+      const moduleMapping = moduleMappings[importedName];
+      if (!moduleMapping) {
+        module = moduleName;
+      } else {
+        module = moduleMapping[0];
+        importedName = moduleMapping[1] || 'default';
+      }
+
+      newImports[module] = newImports[module] || [];
+      newImports[module].push({ localName, importedName });
+    });
+
+    const lines = Object.keys(newImports).map(module => {
+      const newModuleImport = newImports[module];
+
+      const defaultImport = newModuleImport
+        .filter(it => it.importedName === 'default')
+        .map(it => it.localName);
+
+      const namedImports = newModuleImport
+        .filter(it => it.importedName !== 'default')
+        .map(it =>
+          it.importedName !== it.localName
+            ? `${it.importedName} as ${it.localName}`
+            : it.importedName
+        )
+        .join(', ');
+
+      const specifiers = defaultImport
+        .concat(namedImports ? `{ ${namedImports} }` : '')
+        .filter(Boolean)
+        .join(', ');
+
+      return `import ${specifiers} from '${module}';`;
+    });
+
+    return fixer.replaceText(node, lines.join('\n'));
+  };
+
+  return fix;
+}
+
 module.exports = {
+  buildFix,
   buildMessage,
 };
